cleanup + reorg
